Clarify vintage cleanup documentation
The configuration for the additional service account issuers array got moved during our vintage migration and wrote the cleanup docs taking the old config as reference.  But that config is no longer used in any migrated cluster, and it can lead to confusion like it was in the docs. So I've rewrote the documentation with the new config structure as reference.